Add test leak detection
