+++ go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/siddontang/loggers v1.0.3
 	github.com/sirupsen/logrus v1.9.3
 	github.com/stretchr/testify v1.8.4
+	go.uber.org/goleak v1.2.1
 	golang.org/x/sync v0.4.0
 )
