+++ pkg/checksum/checker_test.go
@@ -4,16 +4,13 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cashapp/spirit/pkg/testutils"
-
-	mysql "github.com/go-sql-driver/mysql"
-	"github.com/sirupsen/logrus"
-
-	"github.com/stretchr/testify/assert"
-
 	"github.com/cashapp/spirit/pkg/dbconn"
 	"github.com/cashapp/spirit/pkg/repl"
 	"github.com/cashapp/spirit/pkg/table"
+	"github.com/cashapp/spirit/pkg/testutils"
+	mysql "github.com/go-sql-driver/mysql"
+	"github.com/sirupsen/logrus"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestBasicChecksum(t *testing.T) {
